+++ retroarch.h
@@ -209,7 +209,7 @@ void retroarch_fail(int error_code, const char *error);
 
 bool should_quit_on_close(void);
 
-uint16_t retroarch_get_flags(void);
+uint32_t retroarch_get_flags(void);
 
 RETRO_END_DECLS
